+++ packages/lib/box-forge/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.85](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-box-forge@0.0.84...@opendaw/lib-box-forge@0.0.85) (2026-05-04)
+
+**Note:** Version bump only for package @opendaw/lib-box-forge
+
 ## [0.0.84](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-box-forge@0.0.83...@opendaw/lib-box-forge@0.0.84) (2026-04-22)
 
 **Note:** Version bump only for package @opendaw/lib-box-forge