+++ packages/lib/box/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.83](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-box@0.0.82...@opendaw/lib-box@0.0.83) (2026-05-04)
+
+**Note:** Version bump only for package @opendaw/lib-box
+
 ## [0.0.82](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-box@0.0.81...@opendaw/lib-box@0.0.82) (2026-04-22)
 
 **Note:** Version bump only for package @opendaw/lib-box