+++ packages/lib/fusion/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.90](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-fusion@0.0.89...@opendaw/lib-fusion@0.0.90) (2026-05-04)
+
+**Note:** Version bump only for package @opendaw/lib-fusion
+
 ## [0.0.89](https://github.com/andremichelle/openDAW/compare/@opendaw/lib-fusion@0.0.88...@opendaw/lib-fusion@0.0.89) (2026-04-22)
 
 **Note:** Version bump only for package @opendaw/lib-fusion