+++ packages/studio/adapters/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.106](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-adapters@0.0.105...@opendaw/studio-adapters@0.0.106) (2026-05-04)
+
+**Note:** Version bump only for package @opendaw/studio-adapters
+
 ## [0.0.105](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-adapters@0.0.104...@opendaw/studio-adapters@0.0.105) (2026-04-30)
 
 **Note:** Version bump only for package @opendaw/studio-adapters