+++ packages/studio/boxes/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.88](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-boxes@0.0.87...@opendaw/studio-boxes@0.0.88) (2026-05-04)
+
+**Note:** Version bump only for package @opendaw/studio-boxes
+
 ## [0.0.87](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-boxes@0.0.86...@opendaw/studio-boxes@0.0.87) (2026-04-22)
 
 **Note:** Version bump only for package @opendaw/studio-boxes