+++ packages/studio/p2p/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.8](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-p2p@0.0.7...@opendaw/studio-p2p@0.0.8) (2026-05-04)
+
+**Note:** Version bump only for package @opendaw/studio-p2p
+
 ## [0.0.7](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-p2p@0.0.6...@opendaw/studio-p2p@0.0.7) (2026-04-30)
 
 **Note:** Version bump only for package @opendaw/studio-p2p